% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{custom_theme}
\alias{custom_theme}
\title{Custom Theme}
\usage{
custom_theme(
  font_size = 14,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14,
  color = "grey85",
  map_void = 1,
  border_size = 2
)
}
\arguments{
\item{color}{}
}
\description{
Custom Theme
}
