% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.R
\name{get_gpas}
\alias{get_gpas}
\title{Get GPAS Data}
\usage{
get_gpas(fiscal_year)
}
\arguments{
\item{fiscal_year}{A quoted character.}
}
\value{
Nothing. Writes to a file gdb.
}
\description{
Get GPAS Data
}
