% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_render.R
\name{get_nps_report}
\alias{get_nps_report}
\title{Run BMP Report}
\usage{
get_nps_report(
  fiscal_year,
  output_dir =
    "T:\\\\FS\\\\NFS\\\\R01\\\\Program\\\\2500WatershedAirMgmt\\\\GIS\\\\WorkSpace\\\\jerickson\\\\Spatial\\\\reports\\\\documents\\\\"
)
}
\arguments{
\item{fiscal_year}{A named list with fiscal_year, ex. list(fiscal_year = '2023').}

\item{output_file}{A file path to download file.}
}
\value{
A rendered HTML report
}
\description{
Run BMP Report
}
