% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{pyfn_env}
\alias{pyfn_env}
\title{Python Function Environment}
\format{
An object of class \code{environment} of length 9.
}
\usage{
pyfn_env
}
\value{
An environment
}
\description{
Python Function Environment
}
\keyword{datasets}
