% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme.R
\docType{class}
\name{Theme-class}
\alias{Theme}
\alias{Theme-class}
\title{A Reference class to represent the theme on the plots.}
\description{
A Reference class to represent the theme on the plots.
}
\section{Fields}{

\describe{
\item{\code{font}}{Character representing the font style. The default value is \code{"Helvetica"}.}

\item{\code{x.text.size}}{Numeric representing the size of the labels on the x-axis. Default value is 12.}

\item{\code{y.text.size}}{Numeric representing the size of the labels on the y-axis. Default value is 12.}
}}
\examples{
theme <- Theme$new()
theme$x.text.size <- 14

}
\author{
Zsolt Kovari
}

