% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_render.R
\name{get_bmp_report}
\alias{get_bmp_report}
\title{Run BMP Report}
\usage{
get_bmp_report()
}
\arguments{
\item{fiscal_year}{A quoted character.}
}
\value{
A rendered HTML report
}
\description{
Run BMP Report
}
